refactor day24 to operate on solutions directly
This both reduces a _ton_ of type-wrangling, and also means that we
can take advantage of an optimization: unlike in previous iterations,
we don't exhaustively enumerate all possible solutions, just those
which differ for the footwell.

Consequently, we've finally actually optimized this solution:

$ hyperfine 'target/release/day24 --part2'
Wed 12 May 2021 05:54:29 PM CEST
Benchmark #1: target/release/day24 --part2
  Time (mean ± σ):     678.4 ms ±   3.4 ms    [User: 677.8 ms, System: 0.4 ms]
  Range (min … max):   674.5 ms … 684.4 ms    10 runs
