+++ Cargo.toml
@@ -3,5 +3,4 @@ members = [
     "day01", "day02", "day03", "day04", "day05", "day06", "day07", "day08", "day09", "day10",
     "day11", "day12", "day13", "day14", "day15", "day16", "day17", "day18", "day19", "day20",
     "day21", "day22", "day23", "day24", "day25",
-    "day24/bounded-permutation-generation",
 ]
