--- day24/bounded-permutation-generation/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "bounded-permutation-generation"
-version = "0.1.0"
-authors = ["Peter Goodspeed-Niklaus <peter.r.goodspeedniklaus@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = "1.0.24"
