+++ day24/src/packing_list.rs
@@ -1,11 +1,9 @@
-use std::convert::TryInto;
-
 use crate::{Compartment, Configurator, Package};
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct PackingList<'a> {
-    pub(crate) configurator: &'a Configurator,
-    pub(crate) compartments: Vec<Compartment>,
+    pub(crate) configurator: &'a Configurator<'a>,
+    pub(crate) compartments: Vec<Option<Compartment>>,
 }
 
 impl<'a> PackingList<'a> {
@@ -14,7 +12,9 @@ impl<'a> PackingList<'a> {
             .packages
             .iter()
             .zip(self.compartments.iter())
-            .filter_map(move |(package, present_in)| (*present_in == want).then(|| *package))
+            .filter_map(move |(package, assignment)| {
+                assignment.and_then(|have| (have == want).then(|| *package))
+            })
     }
 
     /// Quantum Entanglement of the footwell of this sleigh.
@@ -26,61 +26,5 @@ impl<'a> PackingList<'a> {
 
     pub fn weight(&self, compartment: Compartment) -> Package {
         self.packages_in(compartment).sum()
-    }
-}
-
-pub(crate) fn compartments_from_groups(
-    footwell: &[Package],
-    left_saddle: &[Package],
-    right_saddle: &[Package],
-    trunk: &[Package],
-) -> Vec<Compartment> {
-    let mut compartments =
-        Vec::with_capacity(footwell.len() + left_saddle.len() + right_saddle.len() + trunk.len());
-
-    let mut groups = [
-        footwell.to_vec(),
-        left_saddle.to_vec(),
-        right_saddle.to_vec(),
-        trunk.to_vec(),
-    ];
-    groups.iter_mut().for_each(|group| group.sort_unstable());
-
-    while let Some(idx_of_highest) = groups
-        .iter()
-        .map(|group| group.last())
-        .enumerate()
-        .filter_map(|(idx, maybe_top)| maybe_top.map(|top| (top, idx)))
-        .max()
-        .map(|(_, idx)| idx)
-    {
-        groups[idx_of_highest].pop();
-        compartments.push(idx_of_highest.try_into().unwrap());
-    }
-
-    compartments
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_compartments_from_groups() {
-        assert_eq!(
-            compartments_from_groups(&[11, 9], &[10, 8, 2], &[7, 5, 4, 3, 1], &[]),
-            vec![
-                Compartment::Footwell,    // 11
-                Compartment::LeftSaddle,  // 10
-                Compartment::Footwell,    // 9
-                Compartment::LeftSaddle,  // 8
-                Compartment::RightSaddle, // 7
-                Compartment::RightSaddle, // 5
-                Compartment::RightSaddle, // 4
-                Compartment::RightSaddle, // 3
-                Compartment::LeftSaddle,  // 2
-                Compartment::RightSaddle, // 1
-            ]
-        );
     }
 }
